I added the logic in the newGame.java file and added the onClick for each of the buttons. It currently highlights it green and disables the key if correct, and red with the key disabled if incorrect. I also added a popup when the game is won or lost. The logic is not working correctly for that yet. Play again restarts the whole app and exit just closes the popup.
My for loop is currently messed up, I know it is something simple but I can not figure it out atm. It currently just checks the first letter. A symptom of the break I think but without the break it progesses way too fast. Maybe I need to a different iterator.