I fixed the logic. Used a boolean to track if a letter is used and then made the game do somthing depending on that logic after the for loop.
Now the current issue with the logic is words with the same character used twice will not complete the game.

Still needs the logic to add the word so far somewhere on the screen.
And probably the logic to type out the entire word and compare that.

We still need to separate the arms into left arm and right arm, and probably switch the best practices for hard encoded strings in the code.